+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🐞 Fixed
+- Improved recovery from missed ringing events. [#1302](https://github.com/GetStream/stream-video-swift/pull/1302)
+
 ### 🔄 Changed
 
 ### 🐞 Fixed